requirements.txt
-i https://pypi.org/simple
aws-lambda-powertools[parser]==2.8.0
boto3==1.26.69
botocore==1.29.69 ; python_version >= '3.7'
ecdsa==0.18.0 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
fsspec==2023.1.0 ; python_version >= '3.7'
jmespath==1.0.1 ; python_version >= '3.7'
pyasn1==0.5.0rc2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pydantic==1.10.4
pynamodb==5.4.0
python-datauri==1.1.0
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.21.1
python-jose==3.3.0
python-ulid==1.1.0
pyyaml==6.0
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
s3fs==0.4.2
s3transfer==0.6.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==4.4.0 ; python_version >= '3.7'
urllib3==1.26.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
